feat(task): add task management components including AddTaskButton, DraggableTask, and TaskDetailsDialog
