+++ src/components/task/index.ts
@@ -1,0 +1,4 @@
+export { TaskAddDialog, type TaskAddDialogProps, type EmptyStateTaskInput } from "./task-add-dialog";
+export { TaskDetailsDialog } from "./task-details-dialog";
+export { AddTaskButton } from "./add-task-button";
+export { DraggableTask } from "./draggable-task";
